chore: reorganize server module similar to client
