+++ example/example/server.py
@@ -1,13 +1,12 @@
 from typing import Any, Callable
 
-from connecpy import context
-from connecpy.interceptor import AsyncConnecpyServerInterceptor
+from connecpy.server import ServiceContext, ServerInterceptor
 
 from . import haberdasher_connecpy
 from .service import HaberdasherService
 
 
-class MyInterceptor(AsyncConnecpyServerInterceptor):
+class MyInterceptor(ServerInterceptor):
     def __init__(self, msg):
         self._msg = msg
 
@@ -15,7 +14,7 @@ class MyInterceptor(AsyncConnecpyServerInterceptor):
         self,
         method: Callable,
         request: Any,
-        ctx: context.ServiceContext,
+        ctx: ServiceContext,
         method_name: str,
     ) -> Any:
         print("intercepting " + method_name + " with " + self._msg)
