+++ src/connecpy/server.py
@@ -1,15 +1,11 @@
-class ConnecpyServer:
-    """
-    Represents a Connecpy server that handles incoming requests and dispatches them to the appropriate endpoints.
-    """
+from ._server_async import ConnecpyASGIApplication
+from ._server_shared import Endpoint, ServerInterceptor, ServiceContext
+from ._server_sync import ConnecpyWSGIApplication
 
-    def __init__(self):
-        self._endpoints = {}
-        self._prefix = ""
-
-    @property
-    def prefix(self):
-        """
-        Represents the prefix used for routing requests to endpoints.
-        """
-        return self._prefix
+__all__ = [
+    "ConnecpyASGIApplication",
+    "ConnecpyWSGIApplication",
+    "Endpoint",
+    "ServerInterceptor",
+    "ServiceContext",
+]
